refactor: NewCardTopBar -> PaymentsDefaultTopBar 수정
- 다른 Screen에서도 기본 TopBar로 적용되게 수정
