+++ app/src/main/java/nextstep/payments/ui/component/topbar/PaymentsDefaultTopBar.kt
@@ -1,4 +1,4 @@
-package nextstep.payments.ui.component
+package nextstep.payments.ui.component.topbar
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
@@ -15,13 +15,14 @@ import nextstep.payments.R
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun NewCardTopBar(
+fun PaymentsDefaultTopBar(
+    title: String,
     onBackClick: () -> Unit,
     onSaveClick: () -> Unit,
     modifier: Modifier = Modifier,
 ) {
     TopAppBar(
-        title = { Text("카드 추가") },
+        title = { Text(title) },
         navigationIcon = {
             IconButton(onClick = { onBackClick() }) {
                 Icon(
